+++ templates/javascript/jsdoc/components-schemas.jsdoc.js
@@ -20,6 +20,7 @@ function componentsSchemasJsdoc(doc) {
       doc,
       schemaTypes,
       schema,
+      inTypeDef: true,
       isTypeDef: true,
     });